feat/split-responsibilities-base-target-adapter-class: refactor remote and local, place in separate modules
