+++ snowshu/adapters/target_adapters/base_remote_target_adapter.py
@@ -1,0 +1,4 @@
+from snowshu.adapters.target_adapters.base_target_adapter import BaseTargetAdapter
+
+class BaseRemoteTargetAdapter(BaseTargetAdapter):
+    """Base class for all remote target adapters eg. Snowflake, BigQuery etc."""
