+++ snowshu/core/graph_set_runner.py
@@ -15,7 +15,8 @@ import networkx as nx
 from snowshu.core.models import Relation
 from snowshu.adapters.base_sql_adapter import BaseSQLAdapter
 from snowshu.adapters.source_adapters.base_source_adapter import BaseSourceAdapter
-from snowshu.adapters.target_adapters.base_target_adapter import BaseRemoteTargetAdapter, BaseLocalTargetAdapter
+from snowshu.adapters.target_adapters.base_local_target_adapter import BaseLocalTargetAdapter
+from snowshu.adapters.target_adapters.base_remote_target_adapter import BaseRemoteTargetAdapter
 from snowshu.core import utils
 from snowshu.core.compile import RuntimeSourceCompiler
 from snowshu.logger import duration
